+++ examples/numpy_style.py
@@ -95,7 +95,7 @@ class ExampleClass:
     @property
     def readonly_property(self):
         """str: Read-only property documentation."""
-        return "readonly_property"
+        return "readonly property"
 
     @property
     def readwrite_property(self) -> List[int]: